assets/02-powerup/pipeline/scripts/requirements.txt
# RAG Pipeline Dependencies
openai>=1.0.0
supabase>=2.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
python-docx>=1.0.0
pdfplumber>=0.10.0
